inscriure infant excursio, mostrar infants inscrits en una excursio

+++ app/Http/Controllers/ExcursioController.php
@@ -6,7 +6,6 @@ use App\Models\Grup;
 use App\Models\Excursio;
 use App\Models\TipoExcursio;
 use Illuminate\Http\Request;
-use Monolog\Handler\IFTTTHandler;
 use App\Http\Requests\ExcursioRequest;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\ExcursioEditRequest;
@@ -15,39 +14,32 @@ class ExcursioController extends Controller
 {
 
     /**
-     * function will apply the auth middleware to all
-     * methods in the ProjectController except the index and show methods
+     * This function is called when the class is instantiated. It tells the class to run the
+     * middleware function called auth
      */
     public function __construct() {
-        //$this->middleware('auth')->only('create', 'edit'); Si volem fer que pels metodes create i edit l'usuari hagi d'estar autentificat
-        $this->middleware('auth'); // S'aplicara el middleware auth (l'usuari haura d'estar autenitificat) a tots els metodes del ProjectController excepte index i show
+        $this->middleware('auth');
     }
 
 
+
+    
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
+     * A function that returns a view with the excursions, groups and types of excursions.
+     * 
+     * @param Request request The request object.
+     * @return view excursions.index
      */
     public function index(Request $request)
     {
         
         $grups = Grup::get();
         $tiposExcursions = TipoExcursio::get();
-
-
-        // $dataActual = date('Y-m-d');
-        // $dataActualArray = explode("-", $dataActual);
-        // $anyActual = $dataActualArray[0];
-        // $mesActual = $dataActualArray[1];
-        // $diaActual = $dataActualArray[2];
-
         
+        // Si no han buscat res pel buscador mostro totes les excursions
         if ($request->search == '' && $request->grup == NULL && $request->tipoExcursio == NULL) {
 
             $excursions = Excursio::get();
-
-            // $excursions = Excursio::whereDate('created_at','>',date('Y-m-d'));
 
             return view('excursions.index', [
                 'excursions' => $excursions,
@@ -55,6 +47,7 @@ class ExcursioController extends Controller
                 'tiposExcursions' => $tiposExcursions
             ]);
 
+        // Si no filtro segons els inputs de l'usuari
         } else {
 
             $nom = $request->get('search');
@@ -77,8 +70,15 @@ class ExcursioController extends Controller
         }    
     }
 
+    /**
+     * It takes the input from the search bar, and searches the database for any matches
+     * 
+     * @param Request request The request object.
+     * @return data matches from database
+     */
     public function excursions(Request $request) {
        
+        // Pregunto a la base de dades si exiteix una excursió amb el nom escrit per l'usuari
         $querys = Excursio::where('nom', 'LIKE', '%'. $request->term . "%")->get();
 
         $data = [];
@@ -92,13 +92,16 @@ class ExcursioController extends Controller
         return $data;
     }
 
+
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
+     * Show the form for creating a new excursio.
+     * @return view excursions.create
      */
     public function create()
     {
+        // Primer comprovo si l'usuari esta autoritzat i després mostro el formulari per crear una nova excursió
+        $this->authorize('monitor', Persona::class);
+
         $grups = Grup::get();
         
         $tiposExcursions = TipoExcursio::get();
@@ -113,13 +116,16 @@ class ExcursioController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param  App\Http\Requests\ExcursioRequest  $request validates the data entered by the user
+     * @return route excursions.index
      */
-    public function store(ExcursioRequest $request)
+    public function store(ExcursioRequest $request) 
     {
 
+        // Valido que l'usuari estigui autoritzat en crear una nova excursió
+        $this->authorize('monitor', Persona::class);
 
+        // Comprovo que com a minim el monitor hagi seleccionat un grup per anar d'excursió
         if (!isset($request["grups"])) {
             return redirect()->route('excursions.create')
                 ->with(['errorsExcursio' => "No has seleccionat cap grup perquè vagi d'excursió!",
@@ -131,6 +137,7 @@ class ExcursioController extends Controller
                         'descripcio' => $request->descripcio]);
         }
 
+        // Comprovo que si la data d'inici de l'excursió és igual a la data de fi l'hora de sortida sigui més gran que l'hora d'arribada
         if ($request->data_inici == $request->data_fi) {
 
             $horaIniciFormulari = explode(':',$request->hora_inici);
@@ -168,6 +175,7 @@ class ExcursioController extends Controller
             }
         }
 
+        // Emmagatzemo la imatge i l'autorització a la carpeta storage\app\public
         $imatge = $request->file('imatge')->store('public');
         $imatgeRuta = explode('/', $imatge);
         $imatgeNom = $imatgeRuta[1];
@@ -177,7 +185,7 @@ class ExcursioController extends Controller
         $autoritzacioNom = $autoritzacioRuta[1];
 
     
-
+        // Emmagatzemo la nova excursió a la base de dades
         $excursio = Excursio::create([
             'tipo_excursio_id' => $request->tipo_excursio_id,
             'nom' => $request->nom, 
@@ -192,7 +200,7 @@ class ExcursioController extends Controller
             'long' => $request->long
         ]);
         
-
+        // Emmagatzemo les relacions de la taula excursions_grups
         if (count($request["grups"])>0) {
 
             foreach ($request["grups"] as $grup_id) {
@@ -202,20 +210,17 @@ class ExcursioController extends Controller
         }
 
         return redirect()->route('excursions.index')->with('status', 'Excursió creada amb èxit.');
-
-        //$request->file('imatge')->store('public'); //storage/app/public
-
-        //$imatge = $request->file('imatge') PER OBTENIR LA IMATGE
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  
-     * @return \Illuminate\Http\Response
+     * @param  App\Models\Excursio $excursio instance of the model Excursio
+     * @return view excursions.show
      */
     public function show(Excursio $excursio)
     {
+        // Dono format a les dates i retorno la vista passant per paràmetre la instància del model Excursio
         $dataIniciBD = explode(' ',$excursio->data_inici);
         $dataIniciArray = explode('-',$dataIniciBD[0]);
         $dataInici = $dataIniciArray[2] . '/' . $dataIniciArray[1] . "/" . $dataIniciArray[0];
@@ -242,11 +247,15 @@ class ExcursioController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param  App\Models\Excursio $excursio instance of the model Excursio
+     * @return view excursions.edit 
      */
     public function edit(Excursio $excursio)
     {
+        // Comprovo que l'usuari estigui autoritzat per editar una excursió
+        $this->authorize('monitor', Persona::class);
+
+        // Dono format a les dates, obtinc els grups que van d'excursió i el tipos d'excursió
         $dataIniciBD = explode(' ',$excursio->data_inici);
         $dataInici = $dataIniciBD[0];
         $horaInici = $dataIniciBD[1];
@@ -263,6 +272,7 @@ class ExcursioController extends Controller
             $grupsExcursio[] = $grupExcursio->grup_id;
         }
 
+        // Retorno el formulari amb la vista perquè l'usuari pugui editar les dades de l'excursió
         return view('excursions.edit', [
             'excursio' => $excursio,
             'tiposExcursions' => $tiposExcursions,
@@ -278,12 +288,16 @@ class ExcursioController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param  App\Models\Excursio  $excursio instance of the model Excursio
+     * @param  App\Http\Requests\ExcursioEditRequest $request validates the data entered by the user
+     * @return view excursions.index 
      */
     public function update(Excursio $excursio, ExcursioEditRequest $request)
     {
+        // Comprovo que l'usuari estigui autoritzat per editar una excursió
+        $this->authorize('monitor', Persona::class);
+
+        // Comprovo que com a minim el monitor hagi seleccionat un grup per anar d'excursió
         if (!isset($request["grups"])) {
             return redirect()->route('excursions.edit', $excursio)
                 ->with(['errorsExcursio' => "No has seleccionat cap grup perquè vagi d'excursió!",
@@ -295,6 +309,7 @@ class ExcursioController extends Controller
                         'descripcio' => $request->descripcio]);
         }
 
+        // Comprovo que si la data d'inici de l'excursió és igual a la data de fi l'hora de sortida sigui més gran que l'hora d'arribada
         if ($request->data_inici == $request->data_fi) {
 
             $horaIniciFormulari = explode(':',$request->hora_inici);
@@ -332,6 +347,7 @@ class ExcursioController extends Controller
             }
         }
 
+        // Si el monitor afegeix una nova imatge elimino l'antiga i emmagatzemo la nova
         if ($request->hasFile('imatge')) {
             Storage::delete('public/'.$excursio->imatge);
 
@@ -341,6 +357,7 @@ class ExcursioController extends Controller
             $excursio->imatge = $imatgeNom;
         }
         
+        // Si el monitor afegeix una nova autorització elimino l'antiga i emmagatzemo la nova
         if ($request->hasFile('autoritzacio')) {
             Storage::delete('public/'.$excursio->autoritzacio);
 
@@ -350,7 +367,7 @@ class ExcursioController extends Controller
             $excursio->autoritzacio = $autoritzacioNom;
         }    
 
-
+        // Actualitzo les dades de l'excursió
         $excursio->tipo_excursio_id = $request->tipo_excursio_id;
         $excursio->nom = $request->nom;
         $excursio->preu = $request->preu;
@@ -362,34 +379,43 @@ class ExcursioController extends Controller
         $excursio->long = $request->long;
         $excursio->save();
 
+        // Obtinc els grups que van d'excursió
         $grupsExcursio = [];
 
         foreach ($request->grups as $grup_id) {
             $grupsExcursio[] = $grup_id;
         }
 
-      
+        // Actualitzo les relacions de la taula excursions_grups
         $excursio->grups()->sync($grupsExcursio);
            
 
-
+        // Redirigeixo a l'usuari a la ruta excursions.index
         return redirect()->route('excursions.index')->with('status', 'Excursió actualitzada amb èxit.');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param  App\Models\Excursio  $excursio instance of the model Excursio
+     * @return route excursions.index
      */
     public function destroy(Excursio $excursio)
     {
+        // Comprovo que l'usuari estigui autoritzat per editar una excursió
+        $this->authorize('monitor', Persona::class);
+
+        // Elimino la imatge i l'autorització de l'excursió
         Storage::delete('public/'.$excursio->imatge);
         Storage::delete('public/'.$excursio->autoritzacio);
 
+        // Elimino les relacions de la taula excursions_grups
         $excursio->grups()->detach();
+
+        // Elimino l'excursió
         $excursio->delete();
 
+        // Redirigeixo a l'usuari a la ruta excursions.index
         return redirect()->route('excursions.index')->with('status','Excursió eliminada amb èxit.');
     }
 }
