+++ app/Http/Controllers/InfantController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+
 use App\Models\Curs;
 use App\Models\Grup;
 use App\Models\Infant;
@@ -9,34 +10,44 @@ use App\Models\Persona;
 use App\Models\Poblacio;
 use App\Models\InfantSalut;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 use App\Http\Requests\InfantRequest;
 
 class InfantController extends Controller
 {
+    /**
+     * This function is called when the class is instantiated. It tells the class to run the
+     * middleware function called auth
+     */
     public function __construct() {
-        //$this->middleware('auth')->only('create', 'edit'); Si volem fer que pels metodes create i edit l'usuari hagi d'estar autentificat
-        $this->middleware('auth')->except('index','show'); // S'aplicara el middleware auth (l'usuari haura d'estar autenitificat) a tots els metodes del ProjectController excepte index i show
+        $this->middleware('auth'); 
     }
 
-
+    
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
+     * If the search field is empty, it returns all the groups. If it's not empty, it returns the
+     * groups that contain the searched text
+     * 
+     * @param Request request The request object
+     * @return view infants.index
      */
     public function index(Request $request)
     {
+        // Comprovo que l'usuari estigui autoritzat a veure el llistat d'infants
+        $this->authorize('monitor', Persona::class);
+
+
         $text = $request->search;
         
-
+        // Si el buscador està buit obtinc tots els grups i mostro el llistat d'infants separats per grups
         if ($text == '') {
 
             $grups = Grup::get();
             return view('infants.index', compact('grups'));
 
+        // Si no busco el nom o cognoms introduït per l'usuari a la base de dades    
         } else {
 
+            // Si conté un espai busco nom i cognoms
             if (strpos($text, ' ') !== false) {
             
                 $split = explode(" ", $text);
@@ -45,55 +56,46 @@ class InfantController extends Controller
     
                 $infants = Persona::where('nom', 'LIKE', "%" . $nom . "%")->where('cognoms', 'LIKE', "%" . $cognoms . "%")->whereNotNull('targeta_sanitaria')->get();
                 
-    
+            // Si o busco el nom o el cognom
             } else {
                 $infants = Persona::where('nom', 'LIKE', "%" . $text . "%")->orwhere('cognoms', 'LIKE',  "%" . $text. "%")->whereNotNull('targeta_sanitaria')->get();
             }
 
+            // Obtinc el grup o els grups al qual pertany l'infant
             $grups = [];
-
             foreach ($infants as $persona) {
                 if ($persona->targeta_sanitaria != NULL) {
-                    $grups[] = $persona->infant->grup;
+                    if (!in_array($persona->infant->grup, $grups)) {
+                        $grups[] = $persona->infant->grup;
+                    }
                 }
-                
             }
 
+            // Si l'array de grups esta buit significa que no ha trobat cap infant així que li mostro un missatge a l'usuari
             if (empty($grups)) {
 
                 return view('infants.index', compact('grups','infants'))->with('statusSearch','ko');
             }
 
+            // Si no retorno la vista amb els grups i els infants
             return view('infants.index', compact('grups','infants'));
         }    
-
-        // $id = 6;
-        // $grup = Grup::find($id);
-        // $espurnes = $grup->infants;
-
-        // foreach ($espurnes as $espurna) {
-        //     return $espurna->persona['nom'];
-        // }
-        // $infantsEspurnes = $espurnes->persona;
-
-        // return $infantsEspurnes;
-
-        // $id = 1;
-        // $infant = Infant::find($id);
-
-        //return $infant->grup;
-
-        // return view('infants.index', [
-        //     'espurnes' => $espurnes
-        // ]);
-
-        
-    
-
     }
 
+
+    /**
+     * It searches for a person in the database and returns the first 5 results.
+     * 
+     * @param Request request The request object.
+     * 
+     * @return data matches from database
+     */
     public function infants(Request $request) {
 
+        // Comprovo que l'usuari estigui autoritzat
+        $this->authorize('monitor', Persona::class);
+
+        // Realitzo la busqueda a la base de dades segons el text introduït per l'usuari
         if (strpos($request->term, ' ') !== false) {
             
             $split = explode(" ", $request->term);
@@ -102,15 +104,13 @@ class InfantController extends Controller
 
             $querys = Persona::where('nom', 'LIKE', '%'. $nom . "%")->where('cognoms', 'LIKE', '%' . $cognoms . '%')->whereNotNull('targeta_sanitaria')->get();
             
-
         } else {
             
             $querys = Persona::where('nom', 'LIKE', '%'. $request->term . "%")->orwhere('cognoms', 'LIKE',  "%" . $request->term . "%")->get();
 
-
         }
-        
 
+        // Afeigeixo el nom i cognoms de l'infant a l'array data
         $data = [];
 
         foreach ($querys as $query) {
@@ -129,17 +129,30 @@ class InfantController extends Controller
         return $data;
     }
 
-    // MOSTRAR UN INFANT
+    /**
+     * Display the specified resource.
+     *
+     * @param  App\Models\Persona $persona instance of the model Persona
+     * @return view infants.show
+     */
     public function show(Persona $persona) {
+
+        // Primer comprovo si l'usuari esta autoritzat i després mostro les dades de l'infant
+        $this->authorize('show', $persona);
+
         return view('infants.show', [
             'persona' => $persona,
         ]);
     }
 
     /**
-     * Per mostrar la vista
+     * Show the form for creating a new infant.
+     * @return view infants.create
      */
     public function create() {
+
+        // Primer comprovo si l'usuari esta autoritzat i després mostro el formulari per crear un nou infant
+        $this->authorize('create', Persona::class);
 
         $poblacions = Poblacio::get();
 
@@ -156,117 +169,68 @@ class InfantController extends Controller
     }
 
     /**
-     * Per processar el formulari
+     * Store a newly created resource in storage.
+     *
+     * @param  App\Http\Requests\InfantRequest  $request validates the data entered by the user
+     * @return route tutors.create
      */
     public function store (InfantRequest $request) {
 
-        // NO FA FALTA PERQUE VALIDEM LES DADES AL InfantRequest -------------------------------------------------------------- BORRAAAAR!!
-        // $formulari = request()->validate([
-        //     'nom' => 'required',
-        //     'cognoms' => 'required',
-        //     'email' => 'required|email',
-        //     'telefon' => 'required|min:9|max:9',
-        //     'data_naixement' => 'required',
-        //     'curs' => 'required',
-        //     'grup' => 'required',
-        //     'targeta_sanitaria' => 'required',
-        //     'carrer' => 'required',
-        //     'poblacio' => 'required',
-        //     'codi_postal' => 'required|min:5|max:5',
-        //     'dni' => '',
-        //     'alergies' => 'required',
-        //     'alergia' => ''
-        // ], [
-        //     'nom.required' => 'Introdueix el nom!!!!'
-        // ]); -------------------------------------------------------------------------------------------- BORRAR FINS AQUI
-
-        // $existeixRegistre = DB::table('persones')->where('targeta_sanitaria', request('targeta_sanitaria'))->exists(); HO FAIG PER COMPROVAR SI LA TARGETA SANITARIA EXISTEIX
-
-        // if ($existeixRegistre == false) {
-
-            
-            Persona::create($request->validated());
-           
-            
-
-            // ES EL MATEIX QUE FER AIXO -------------------------------------------------------------- BORRAAAAR!!
-            // if (is_null(request('dni'))) { // $formulari['dni']
-            //     Persona::create([
-            //         'poblacio_id' => request('poblacio'),
-            //         'nom' => request('nom'), 
-            //         'cognoms'=> request('cognoms'),
-            //         'email'=> request('email'),
-            //         'telefon'=> request('telefon'),
-            //         'carrer'=> request('carrer'),
-            //         'codi_postal' => request('codi_postal'),
-            //         'data_naixement' => request('data_naixement'),
-            //         'targeta_sanitaria' => request('targeta_sanitaria')
-            //     ]);
-                
-            // } else {
+        // Comprovo que l'usuari estigui autoritzat
+        $this->authorize('create', Persona::class);
+        
+        // Creo una nova instància del model Persona i emmagatzemo les dades a la base de dades
+        $persona = Persona::create($request->validated());
     
-            //     Persona::create([
-            //         'poblacio_id' => request('poblacio'),
-            //         'nom' => request('nom'), 
-            //         'cognoms'=> request('cognoms'),
-            //         'email'=> request('email'),
-            //         'telefon'=> request('telefon'),
-            //         'carrer'=> request('carrer'),
-            //         'codi_postal' => request('codi_postal'),
-            //         'data_naixement' => request('data_naixement'),
-            //         'dni' => request('dni'),
-            //         'targeta_sanitaria' => request('targeta_sanitaria')
-            //     ]);
-                
-            // } -------------------------------------------------------------------------------------------- BORRAR FINS AQUI
-    
-            $personaId = Persona::latest()->first();
-    
-            Infant::create([
-                'persona_id' => $personaId['persona_id'],
-                'grup_id' => request('grup'),
-                'curs_id' => request('curs')
+        // Creo una nova instància del model Infant i emmagatzemo les dades a la base de dades
+        $infant = Infant::create([
+            'persona_id' => $persona->persona_id,
+            'grup_id' => request('grup'),
+            'curs_id' => request('curs')
+        ]);
+
+        // Comprovo si l'infant té al·lèergies i creo una nova instància del model InfantSalut i emmagatzemo les dades a la base de dades
+        if($request['alergies'] == 1) {
+
+            InfantSalut::create([
+                'infant_id' => $infant->infant_id,
+                'alergies' => request('alergies'),
+                'alergia' => request('alergia')
             ]);
 
-            $infantId = Infant::latest()->first();
+        } else {
 
-            if($request['alergies'] == 1) {
-
-                InfantSalut::create([
-                    'infant_id' => $infantId['infant_id'],
-                    'alergies' => request('alergies'),
-                    'alergia' => request('alergia')
-                ]);
-
-            } else {
-
-                InfantSalut::create([
-                    'infant_id' => $infantId['infant_id'],
-                    'alergies' => request('alergies')
-                ]);
-            }    
+            InfantSalut::create([
+                'infant_id' => $infant->infant_id,
+                'alergies' => request('alergies')
+            ]);
+        }    
             
+        // Redirigeixo al monitor a la ruta tutors.create perque m'afegeixi un tutor i li passo per paràmetre l'infant
+        return redirect()->route('tutors.create', $persona)->with('statusInfant', 'ok');
 
-            return redirect()->route('tutors.create')->with('statusInfant', 'ok'); // Metode with() per enviar a la vista variables de sessio que duraran fins que l'usuari recarregui la pagina
-
-        // } else {
-            
-        //     return back()->with('message', 'Aquest infant ja exiteix a la base de dades!');
-
-        // }
-
-    
     }
 
 
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  App\Models\Persona $persona instance of the model Persona
+     * @return view infants.edit 
+     */
     public function edit(Persona $persona) {
 
+        // Comprovo que l'usuari estigui autoritzat
+        $this->authorize('update', $persona);
+
+        // Obtinc les poblacions, cursos i grups
         $poblacions = Poblacio::get();
 
         $cursos = Curs::get();
 
         $grups = Grup::get();
 
+        // Mostro la vista infants.edit
         return view('infants.edit', [
             'persona' => $persona,
             'poblacions' => $poblacions,
@@ -275,8 +239,19 @@ class InfantController extends Controller
         ]);
     }
 
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  App\Models\Persona  $persona instance of the model Persona
+     * @param  Request $request data entered by the user
+     * @return view excursions.index 
+     */
     public function update (Persona $persona, Request $request) {
 
+        // Comprovo que l'usuari estigui autoritzat
+        $this->authorize('update', $persona);
+
+        // Valido les dades introduïdes per l'usuari
         $this->validate($request, [
             'nom' => 'required',
             'cognoms' => 'required',
@@ -297,6 +272,7 @@ class InfantController extends Controller
             'dni.unique' => 'El DNI ja està registrat i no es pot repetir.'
         ]);
 
+        // Actualitzo les dades de la instància persona i les emmagatzemo a la base de dades
         $persona->poblacio_id = $request->poblacio_id;
         $persona->nom = $request->nom;
         $persona->cognoms = $request->cognoms;
@@ -309,13 +285,16 @@ class InfantController extends Controller
         $persona->targeta_sanitaria = $request->targeta_sanitaria;
         $persona->save();
 
+        // Obtinc les instàncies infant i infantSalut a partir de persona
         $infant = $persona->infant;
         $infantSalut = $infant->infantSalut;
 
+        // Actualitzo les dades de la instància infant i les emmagatzemo a la base de dades
         $infant->grup_id = $request->grup;
         $infant->curs_id = $request->curs;
         $infant->save();
 
+        // Actualitzo les dades de la instància infantSalut i les emmagatzemo a la base de dades
         if($request['alergies'] == 1) {
 
             $infantSalut->alergies = $request->alergies;
@@ -329,20 +308,36 @@ class InfantController extends Controller
         
         $infantSalut->save();
 
+        // Redirigeixo a l'usuari a la ruta infants.index
         return redirect()->route('infants.index')->with('status', "Infant actualitzat correctament.");
     }
 
 
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  App\Models\Persona  $persona instance of the model Persona
+     * @return route infants.index
+     */
     public function destroy (Persona $persona) {
 
-       $infant = $persona->infant;
-       $infantSalut = $infant->infantSalut;
+        // Comprovo que l'usuari estigui autoritzat
+        $this->authorize('monitor', Persona::class);
 
-       $infant->tutors()->detach();
-       $infantSalut->delete();
-       $infant->delete();
-       $persona->delete();
+        // Obtinc les instàncies infant i infantSalut a partir de persona
+        $infant = $persona->infant;
+        $infantSalut = $infant->infantSalut;
 
-       return redirect()->route('infants.index')->with('statusEliminar','Infant eliminat correctament.');
+        // Elimino les relacions de la taula tutors_infants
+        $infant->tutors()->detach();
+        // Elimino la instància del model InfantSalut
+        $infantSalut->delete();
+        // Elimino la instància del model Infant
+        $infant->delete();
+        // Elimino la instància del model Persona
+        $persona->delete();
+
+        // Redirigeixo a l'usuari a la ruta infants.index
+        return redirect()->route('infants.index')->with('statusEliminar','Infant eliminat correctament.');
     }
 }
